requirements.txt
black
click
isort
mypy
openai
pylint
python-dotenv
